Move _convert_to_time_series, _convert_pulse_to_driving_field to translation file
